Added protobuf bom to gradle libs. Libraries upgrade. Java 17
