+++ protobuff-support/build.gradle
@@ -18,6 +18,8 @@ dependencies {
     testImplementation mn.micronaut.reactor.http.client
     testImplementation mn.micronaut.jackson.databind
     testImplementation libs.jackson.datatype.protobuf
+
+    testRuntimeOnly mn.micronaut.jackson.databind
 }
 
 protobuf {
