+++ test-suite-groovy/build.gradle
@@ -34,15 +34,22 @@ micronaut {
 
 dependencies {
     compileOnly "io.grpc:grpc-stub:$grpcVersion"
-    compileOnly("javax.annotation:javax.annotation-api")
+    compileOnly "javax.annotation:javax.annotation-api"
 
     implementation "io.micronaut:micronaut-validation"
     implementation("io.micronaut:micronaut-runtime")
     implementation(project(':grpc-server-runtime'))
     implementation "io.micronaut.discovery:micronaut-discovery-client"
     implementation "io.grpc:grpc-services"
-    implementation("org.yaml:snakeyaml")
+    implementation "org.yaml:snakeyaml"
+
     runtimeOnly "ch.qos.logback:logback-classic"
+
+    testCompileOnly "javax.annotation:javax.annotation-api"
+
+    testImplementation "org.apache.groovy:groovy"
+
+    testRuntimeOnly "ch.qos.logback:logback-classic"
 }
 
 tasks.withType(GroovyCompile) {
