+++ test-suite-java/build.gradle
@@ -32,18 +32,20 @@ dependencies {
     annotationProcessor "io.micronaut:micronaut-validation"
 
     compileOnly "io.grpc:grpc-stub:$grpcVersion"
-    compileOnly("javax.annotation:javax.annotation-api")
+    compileOnly "javax.annotation:javax.annotation-api"
 
     implementation "io.micronaut:micronaut-validation"
-    implementation("io.micronaut:micronaut-runtime")
-    implementation(project(':grpc-server-runtime'))
+    implementation "io.micronaut:micronaut-runtime"
+    implementation project(':grpc-server-runtime')
     implementation "io.micronaut.discovery:micronaut-discovery-client"
     implementation "io.grpc:grpc-services"
-    implementation("org.yaml:snakeyaml")
+    implementation "org.yaml:snakeyaml"
 
     runtimeOnly "ch.qos.logback:logback-classic"
 
     testAnnotationProcessor "io.micronaut:micronaut-validation"
+
+    testCompileOnly "javax.annotation:javax.annotation-api"
 }
 
 // tag::config[]
