+++ test-suite-kotlin/build.gradle
@@ -46,17 +46,20 @@ dependencies {
     implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:$kotlinxCoroutinesVersion")
     implementation("io.micronaut:micronaut-runtime")
     implementation(project(':grpc-server-runtime'))
-    implementation("org.yaml:snakeyaml")
+    implementation "org.yaml:snakeyaml"
 
     implementation "io.grpc:grpc-services"
 
     // tag::dependencies[]
     implementation("io.grpc:grpc-kotlin-stub:$grpcKotlinVersion")
     compileOnly("io.grpc:grpc-stub:$grpcVersion")
+    compileOnly "javax.annotation:javax.annotation-api"
     // end::dependencies[]
 
     runtimeOnly("ch.qos.logback:logback-classic")
     runtimeOnly("com.fasterxml.jackson.module:jackson-module-kotlin")
+
+    testCompileOnly "javax.annotation:javax.annotation-api"
 }
 
 tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile) {
