+++ test-suite-kotlin/src/main/kotlin/helloworld/GreetingEndpoint.kt
@@ -21,7 +21,6 @@ import jakarta.inject.Singleton
 
 // tag::clazz[]
 @Singleton // <1>
-@Suppress("unused")
 class GreetingEndpoint(val greetingService : GreetingService) : GreeterGrpcKt.GreeterCoroutineImplBase() { // <2>
     override suspend fun sayHello(request: HelloRequest): HelloReply {
         // <3>
